+++ index.js
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Dependencies //////////////////////////////////////////////////////////////////////////////////////////////////////////
-const { Message, Interaction, EmbedBuilder, ButtonBuilder, ButtonStyle } = require("discord.js");
+const { Message, Interaction, EmbedBuilder, ButtonBuilder } = require("discord.js");
 const PaginationBase = require("./lib/PaginationBase");
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Wrapper ///////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -234,6 +234,7 @@ module.exports = class PaginationWrapper {
     *    }],
     *    imageUrl?: String,
     *    color?: 'Default'
+    *       |  'Random'
     *       |  'White'
     *       |  'Aqua'
     *       |  'Green'
