+++ lib/MessagePagination.js
@@ -1,13 +1,13 @@
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Dependencies //////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Dependencies
 const { Message, ActionRowBuilder, EmbedBuilder, ButtonBuilder } = require("discord.js");
-const DisabledButtons = require("../util/DisabledButtons");
-const SelectMenuBuilder = require("../util/SelectMenuBuilder");
-const ProgressBarBuilder = require("../util/ProgressBarBuilder");
-const { FilterBuilder } = require("../util/FilterBuilder");
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Params ////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+const { DisabledButtonCreator } = require("../util/ButtonCreator");
+const { SelectMenuCreator, DisabledSelectMenuCreator } = require("../util/SelectMenuCreator");
+const ProgressBarCreator = require("../util/ProgressBarCreator");
+const { FilterCreator } = require("../util/FilterCreator");
+// Params
 /**
+ * The Message pagination
+ * 
  * @param {Message} message - Discord.js interface
  * @param {EmbedBuilder[]} pageList - An array of the embeds
  * @param {ButtonBuilder[]} buttonList - An array of the buttons
@@ -21,20 +21,19 @@ const { FilterBuilder } = require("../util/FilterBuilder");
  * @param {Boolean} authorIndependent - Only the author can use pagination
  * @param {Boolean} selectMenu - If select menu is enabled
  * @returns {EmbedBuilder[]} The pagination
-*/
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Message pagination ////////////////////////////////////////////////////////////////////////////////////////////////////
+ */
+// Message pagination
 module.exports = MessagePagination = async(message, pageList, buttonList, timeout, replyMessage, autoDelete, privateReply, progressBar, proSlider, proBar, authorIndependent, selectMenu) => {
    try {
       // Set page number
       let pageNumber = 0;
       // Create row
       let row
-      if (selectMenu) {row = await SelectMenuBuilder(pageList.length)}
+      if (selectMenu) {row = await SelectMenuCreator(pageList.length)}
       else {row = new ActionRowBuilder().addComponents(buttonList);}
       // Create embed
       const paginationContent = {
-         embeds: [pageList[pageNumber].setFooter({text: `${progressBar? `${ProgressBarBuilder(pageList.length, pageNumber, proSlider, proBar)}` : `Page ${pageNumber + 1} / ${pageList.length}`}`})],
+         embeds: [pageList[pageNumber].setFooter({text: `${progressBar? `${ProgressBarCreator(pageList.length, pageNumber, proSlider, proBar)}` : `Page ${pageNumber + 1} / ${pageList.length}`}`})],
          components: [row], fetchReply: true
       };
       let pagination;
@@ -46,7 +45,7 @@ module.exports = MessagePagination = async(message, pageList, buttonList, timeou
       }
       // Create collector
       const collector = pagination.createMessageComponentCollector({
-         filter: FilterBuilder({message, buttonList, authorIndependent, selectMenu}),
+         filter: FilterCreator({message, buttonList, authorIndependent, selectMenu}),
          time: timeout
       });
       // Button inputs
@@ -94,7 +93,7 @@ module.exports = MessagePagination = async(message, pageList, buttonList, timeou
             if (!i.deferred) await i.deferUpdate();
             // Edit page after input
             await i.editReply({
-               embeds: [pageList[pageNumber].setFooter({text: `${progressBar ? `${ProgressBarBuilder(pageList.length, pageNumber, proSlider, proBar)}` : `Page ${pageNumber + 1} / ${pageList.length}`}`})],
+               embeds: [pageList[pageNumber].setFooter({text: `${progressBar ? `${ProgressBarCreator(pageList.length, pageNumber, proSlider, proBar)}` : `Page ${pageNumber + 1} / ${pageList.length}`}`})],
                fetchReply: true
             }).catch(error => {return console.log(error)});
          }
@@ -105,7 +104,7 @@ module.exports = MessagePagination = async(message, pageList, buttonList, timeou
             // Edit page after input
             pageNumber = i.values[0] - 1 ;
             await i.editReply({
-               embeds: [pageList[pageNumber].setFooter({text: `${progressBar ? `${ProgressBarBuilder(pageList.length, pageNumber, proSlider, proBar)}` : `Page ${pageNumber + 1} / ${pageList.length}`}`})],
+               embeds: [pageList[pageNumber].setFooter({text: `${progressBar ? `${ProgressBarCreator(pageList.length, pageNumber, proSlider, proBar)}` : `Page ${pageNumber + 1} / ${pageList.length}`}`})],
                fetchReply: true
             }).catch(error => {return console.log(error)});
          }
@@ -122,13 +121,13 @@ module.exports = MessagePagination = async(message, pageList, buttonList, timeou
             // Disable select menu
             if (selectMenu) {
                try {
-                  return pagination.edit({ components: [await SelectMenuBuilder(pageList.length, true)] });
+                  return pagination.edit({ components: [await DisabledSelectMenuCreator(pageList.length)] });
                } catch(error) {}
             }
             // Disable buttons
             else {
                try {
-                  return pagination.edit({ components: [await DisabledButtons(buttonList)] });
+                  return pagination.edit({ components: [await DisabledButtonCreator(buttonList)] });
                } catch (error) {return}
             }
          } catch(error) {return}
