+++ lib/PaginationBase.js
@@ -1,19 +1,17 @@
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Dependencies //////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Dependencies
 const { EmbedBuilder } = require("discord.js");
 const InteractionPagination = require("./InteractionPagination");
 const MessagePagination = require("./MessagePagination");
-const AutoButtonBuilder = require("../util/AutoButtonBuilder");
-const PageBuilder = require("../util/PageBuilder");
-const ButtonBuilder = require("../util/ButtonBuilder");
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Params ////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+const { AutoButtonCreator, ButtonCreator } = require("../util/ButtonCreator");
+const PageCreator = require("../util/PageCreator");
+// Params
 /**
+ * The function that prepares the required data and detects whether its a interaction or message pagination
+ * 
  * @returns {EmbedBuilder[]} The pagination
-*/
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// pagination ////////////////////////////////////////////////////////////////////////////////////////////////////////////
-module.exports = pagination = async({
+ */
+// pagination
+module.exports = PaginationBase = async({
    interaction, message, pageList, buttonList, timeout = 12000,
    replyMessage = false,
    autoDelete = false,
@@ -30,44 +28,44 @@ module.exports = pagination = async({
 }) => {
    try {
       // Checks
-      if (message && interaction) throw new Error("Do not provided both interaction and message only 1 is required");
+      if (message && interaction) throw new Error("PaginationBase ERROR: Do not provided both interaction and message only 1 is required");
       if (!pageList) {
          if (pageBuilderInfo) {
-            pageList = await PageBuilder(pageBuilderInfo);
+            pageList = await PageCreator(pageBuilderInfo);
             console.log(pageList)
          } else {
-            throw new Error("Missing pages");
+            throw new Error("PaginationBase ERROR: Missing pages");
          }
       }
       if (!buttonList) {
          if (buttonBuilderInfo) {
-            buttonList = await ButtonBuilder(buttonBuilderInfo);
+            buttonList = await ButtonCreator(buttonBuilderInfo);
          } else if (autoButton && !buttonBuilderInfo) {
-            buttonList = await AutoButtonBuilder(pageList.length, autoDelButton);
+            buttonList = await AutoButtonCreator(pageList.length, autoDelButton);
          } else {
-            throw new Error("Missing buttons");
+            throw new Error("PaginationBase ERROR: Missing buttons");
          }
       }
-      if (selectMenu && (buttonList || autoButton || autoDelButton)) process.emitWarning("SelectMenu overwrites any button settings, remove all button settings to stop getting this message");
-      if (timeout < 3000) throw new Error("You have set timeout less then 3000ms which is not allowed");
+      if (selectMenu && (buttonList || autoButton || autoDelButton)) process.emitWarning("PaginationBase WARNING: SelectMenu overwrites any button settings, remove all button settings to stop getting this message");
+      if (timeout < 3000) throw new Error("PaginationBase ERROR: You have set timeout less then 3000ms which is not allowed");
       if (progressBar) {
-         if (proSlider.length > 1) throw new Error("You can only use 1 character to represent the progressBar slider");
-         if (proBar.length > 1) throw new Error("You can only use 1 character to the progressBar");
+         if (proSlider.length > 1) throw new Error("PaginationBase ERROR: You can only use 1 character to represent the progressBar slider");
+         if (proBar.length > 1) throw new Error("PaginationBase ERROR: You can only use 1 character to the progressBar");
       }
-      if (buttonList.length < 2) throw new Error("Need provide at least 2 buttons");
+      if (buttonList.length < 2) throw new Error("PaginationBase ERROR: Need provide at least 2 buttons");
       if (buttonList.length > 5) {
-         process.emitWarning("More than 5 buttons have been provided the extras will be removed, remove the extra buttons from the buttonList to stop getting this message");
+         process.emitWarning("PaginationBase WARNING: More than 5 buttons have been provided the extras will be removed, remove the extra buttons from the buttonList to stop getting this message");
          buttonList = buttonList.slice(0, 5);
       }
       for (const button of buttonList) {
-         if (button.style === "LINK") throw new Error("Link buttons are not supported please check what type of buttons you are using");
-         if (button.disabled) throw new Error("You have provided buttons that are disabled these cant be used to turn pages, make sure the buttons you are trying to use are enabled");
+         if (button.style === "LINK") throw new Error("PaginationBase ERROR: Link buttons are not supported please check what type of buttons you are using");
+         if (button.disabled) throw new Error("PaginationBase ERROR: You have provided buttons that are disabled these cant be used to turn pages, make sure the buttons you are trying to use are enabled");
       }
       // Message
       if (typeof message?.author === "object") {
          // Checks
-         if (replyMessage && privateReply) process.emitWarning("The privateReply setting overwrites and disables replyMessage setting");
-         if (!message && !message?.channel) throw new Error("Channel is inaccessible");
+         if (replyMessage && privateReply) process.emitWarning("PaginationBase WARNING: The privateReply setting overwrites and disables replyMessage setting");
+         if (!message && !message?.channel) throw new Error("PaginationBase ERROR: Channel is inaccessible");
          if (pageList.length < 2) {
             if (privateReply) {
                await message.channel.send("The reply has been sent privately");
@@ -92,10 +90,10 @@ module.exports = pagination = async({
          }
          if (interaction.ephemeral) {
             if (buttonList.length === 3 || buttonList.length === 5) {
-               throw new Error("Delete buttons are not supported by embeds with ephemeral enabled");
+               throw new Error("PaginationBase ERROR: Delete buttons are not supported by embeds with ephemeral enabled");
             } 
             if (autoDelete) {
-               throw new Error("Auto delete is not supported by embeds with ephemeral enabled");
+               throw new Error("PaginationBase ERROR: Auto delete is not supported by embeds with ephemeral enabled");
             }
          }
          // Run
@@ -103,7 +101,7 @@ module.exports = pagination = async({
       }
       // Missing interaction and message
       else {
-         throw new Error("Please provide either interaction or message for the pagination to use");
+         throw new Error("PaginationBase ERROR: Please provide either interaction or message for the pagination to use");
       }
    } catch(error) {
       return console.log(`Error occured with ${__filename.split(/[\\/]/).pop().replace(".js","")} ${error}`)
