+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@acegoal07/discordjs-pagination",
-  "version": "1.3.3",
+  "version": "1.3.4",
   "description": "A simple pagination for Discord.js",
   "main": "index",
   "repository": {
