+++ util/FilterCreator.js
@@ -1,12 +1,11 @@
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Params ////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Params
 /**
+ * Crates the filter needed to interpret input for the pagination
  * @returns {Function} The button filter
 */
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Filter Builder ////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Filter Builder
 module.exports = {
-   FilterBuilder({message, interaction, buttonList, authorIndependent, selectMenu}) {
+   FilterCreator({message, interaction, buttonList, authorIndependent, selectMenu}) {
       try {
          // Get author
          let authorID;
