+++ util/PageCreator.js
@@ -1,18 +1,16 @@
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Dependencies //////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Dependencies
 const { EmbedBuilder } = require("discord.js");
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Params ////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Params
 /**
  * Sends back a list of custom pages to use
  * @returns {EmbedBuilder[]}
  */
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Page Builder //////////////////////////////////////////////////////////////////////////////////////////////////////////
-module.exports = Builder = async(pageBuilderInfo) => {
+// Page Builder
+module.exports = PageCreator = async(pageBuilderInfo) => {
    try {
       let pageList = [];
       for (const page of pageBuilderInfo) {
+         if (!page) throw new Error("PageCreator ERROR: blank page info was passed into createPages")
          const embed = new EmbedBuilder();
          if (page.title) embed.setTitle(page.title);
          if (page.description) embed.setDescription(page.description);
